requirements.txt
numpy==1.13
scipy==0.19
pillow==4.0
matplotlib==2.0
h5py==2.7
cython==0.25.2
pandas==0.19
pytest==3.0.7
scikit-image==0.13
jupyter==1.0.0
tqdm==4.11.2
munkres==1.0.12
cython==0.25.2
visdom
opencv-python
imageio